Fix typos, Add new parameter for logger method, revert deleted unit test
