+++ wodles/gcloud/tests/test_tools.py
@@ -4,8 +4,7 @@
 # it and/or modify it under the terms of GPLv2
 
 """Unit tests for gcloud module."""
-
-import logging
+import argparse
 from os.path import join, dirname, realpath
 import sys
 import pytest
@@ -14,15 +13,7 @@ from unittest.mock import patch
 # Local imports
 sys.path.append(join(dirname(realpath(__file__)), '..'))  # noqa: E501 # noqa: E501
 from gcloud import get_script_arguments
-from shared.wazuh_cloud_logger import WazuhCloudLogger
-
-
-@pytest.fixture(scope='module')
-def gcp_logger():
-    return WazuhCloudLogger(
-            logger_name='test_logger'
-            )
-
+from tools import arg_valid_date
 
 def test_get_script_arguments(capsys):
     """Test get_script_arguments shows no messages when the required parameters were provided."""
@@ -48,14 +39,7 @@ def test_get_script_arguments_required(capsys, args):
     assert exception.value.code == 2
 
 
-@pytest.mark.parametrize("log_level, log_method, expected_message", [
-    (logging.INFO, 'info', 'Test info message'),
-    (logging.DEBUG, 'debug', 'Test debug message'),
-    (logging.WARNING, 'warning', 'Test warning message'),
-    (logging.ERROR, 'error', 'Test error message'),
-])
-def test_log_methods(gcp_logger, caplog, log_level, log_method, expected_message):
-    with caplog.at_level(log_level, logger='TestGCPLogStrategy'):
-        with patch.object(gcp_logger, 'logger', logging.getLogger('TestGCPLogStrategy')):
-            getattr(gcp_logger, log_method)(expected_message)
-    assert expected_message in caplog.text
+def test_arg_valid_date():
+    """Test arg_valid_dates raises an error when a date parameter doesn't have a valid format."""
+    with pytest.raises(argparse.ArgumentTypeError):
+        arg_valid_date('invalid_date')
